NAS_MASK/NAS_streamlit_APP/requirements.txt
absl-py==1.4.0
aioice==0.9.0
aiortc==1.5.0
alabaster==0.7.13
altair==5.0.1
antlr4-python3-runtime==4.9.3
attrs==23.1.0
av==10.0.0
Babel==2.12.1
blinker==1.6.2
boto3==1.26.153
botocore==1.29.153
build==0.10.0
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
coloredlogs==15.0.1
contourpy==1.1.0
coverage==5.3.1
cryptography==41.0.1
cycler==0.11.0
decorator==5.1.1
Deprecated==1.2.14
dnspython==2.3.0
docutils==0.17.1
einops==0.3.2
flatbuffers==23.5.26
fonttools==4.40.0
future==0.18.3
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.20.0
google-auth-oauthlib==1.0.0
google-crc32c==1.5.0
grpcio==1.54.2
humanfriendly==10.0
hydra-core==1.3.2
idna==3.4
ifaddr==0.2.0
imagesize==1.4.1
importlib-metadata==6.6.0
imutils==0.5.4
Jinja2==3.1.2
jmespath==1.0.1
json-tricks==3.16.1
jsonschema==4.17.3
kiwisolver==1.4.4
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.1
mdurl==0.1.2
mpmath==1.3.0
numpy==1.23.0
oauthlib==3.2.2
omegaconf==2.3.0
onnx==1.13.0
onnx-simplifier==0.4.31
onnxruntime==1.13.1
opencv-python==4.7.0.72
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
pip-tools==6.13.0
protobuf==3.20.3
psutil==5.9.5
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocotools==2.0.4
pycparser==2.21
pydeck==0.8.1b0
pyDeprecate==0.3.2
pyee==10.0.1
Pygments==2.15.1
pylibsrtp==0.8.0
Pympler==1.0.1
pyOpenSSL==23.2.0
pyparsing==2.4.5
pyproject_hooks==1.0.0
pyrsistent==0.19.3
python-dateutil==2.8.2
pytube==15.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
rapidfuzz==3.1.1
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.4.2
rsa==4.9
s3transfer==0.6.1
scipy==1.10.1
six==1.16.0
smmap==5.0.0
snowballstemmer==2.2.0
Sphinx==4.0.3
sphinx-rtd-theme==1.2.2
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
streamlit==1.23.1
streamlit-camera-input-live==0.2.0
streamlit-webrtc==0.45.1
stringcase==1.2.0
super-gradients==3.1.0
sympy==1.12
tenacity==8.2.2
tensorboard==2.13.0
tensorboard-data-server==0.7.1
termcolor==1.1.0
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==1.13.1
torchmetrics==0.8.0
torchvision==0.14.1
tornado==6.3.2
tqdm==4.65.0
treelib==1.6.1
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.16
validators==0.20.0
Werkzeug==2.3.6
wrapt==1.15.0
zipp==3.15.0
